Add player animations, fix player sorting order
