--- GGJ21/Assets/HideBarriers.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 375584cbe35b9544e80ee34d363bcebc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
