+++ GGJ21/Assets/Prefab/Tilemap Collider.prefab
@@ -185,7 +185,7 @@ Rigidbody2D:
   m_Interpolate: 0
   m_SleepingMode: 1
   m_CollisionDetection: 0
-  m_Constraints: 0
+  m_Constraints: 7
 --- !u!66 &1308706428921659400
 CompositeCollider2D:
   m_ObjectHideFlags: 0
