+++ GGJ21/Assets/Scripts/HideBarriers.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Tilemaps;
+
+public class HideBarriers : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        GetComponent<TilemapRenderer>().enabled = false;
+    }
+}
