+++ GGJ21/Assets/Scripts/HideBarriers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 375584cbe35b9544e80ee34d363bcebc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
