+++ GGJ21/Assets/Scripts/PlayerSprites.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34e508207aa595844b0e9f4c6edd8dd8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
