+++ GGJ21/Assets/Sound/effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b66cd9e24886be4eb82ec6cfa8efd40
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
