+++ GGJ21/Assets/Sound/effects/pickup.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 49db5b863ff84f24d86b3fd605ea727b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
