+++ GGJ21/Assets/Sound/effects/pickup/pickup.ceol.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9e46cc2f1a0c60848a3a0771091a1feb
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
