+++ GGJ21/Assets/Sound/effects/steps.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39a75594aab6e1649bb052de6e14ff92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
