+++ GGJ21/Assets/Sprites/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66dcc8849042e2446a2c33fd388411f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
