+++ GGJ21/Assets/Sprites/Animations/Player.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4cc2d26b3670a9e49b783b469770713d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
