+++ GGJ21/Assets/Sprites/Animations/PlayerDown.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 62e304e3deae7ce4895dc23e3ed536ee
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
