+++ GGJ21/Assets/Sprites/Animations/PlayerIdleDown.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37dcf824eefcc7f449dd95c8f222a594
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
