+++ GGJ21/Assets/Sprites/Animations/PlayerIdleLeft.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd12d46d481e32b4da52e389af4e45d9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
