+++ GGJ21/Assets/Sprites/Animations/PlayerIdleRight.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5eff88c4b7bf53145ae711b1be46bcec
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
