+++ GGJ21/Assets/Sprites/Animations/PlayerIdleUp.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3007d9b24ae8984438caeb205d9ca57a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
