+++ GGJ21/Assets/Sprites/Animations/PlayerLeft.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f35bc1ffdfc0aef45818a7794a574a50
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
