+++ GGJ21/Assets/Sprites/Animations/PlayerRight.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc41929c7dff3be40be055fd73c11aa5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
