+++ GGJ21/Assets/Sprites/Animations/PlayerUp.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a55d92c77194b354fbc399d4b196bea0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
